discussion-hub/requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
ngrok==1.4.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
watchdog==5.0.2
Werkzeug==3.0.4
zipp==3.20.2
python-dotenv
